attempt at fixing historical weather
Had the get one historical reading function never call cb() with an
error, but with an empty record instead.  Problem with most of the async
functions is that if ANY call calls cb() with an error, the whole thing
throws an error - not what we want, in case we are >sometimes< getting
errors.

However, this still didn't help :-(